pyproject.toml
[project]
name = "clustertim"
version = "0.1.0"
description = ""
authors = [
    {name = "Ernesto Cardenas",email = "cernestoc116@hotmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-recaptcha (>=4.1.0,<5.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "qrcode (>=8.2,<9.0)",
    "stripe (>=14.4.1,<15.0.0)",
    "pytz (>=2026.1.post1,<2027.0)",
    "reportlab (>=4.4.10,<5.0.0)",
    "pymupdf (>=1.27.2,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "weasyprint (>=68.1,<69.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "oracledb (>=3.4.2,<4.0.0)",
    "paypal-checkout-serversdk (>=1.0.3,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "whitenoise (>=6.12.0,<7.0.0)",
    "django-import-export (>=4.4.0,<5.0.0)",
    "django-allauth (>=65.15.0,<66.0.0)",
    "django-constance (>=4.3.4,<5.0.0)",
    "django-paypal (>=2.1,<3.0)",
    "django-pwa (>=2.0.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "groq (>=1.1.1,<2.0.0)",
    "openai (>=2.28.0,<3.0.0)",
    "speechrecognition (>=3.15.1,<4.0.0)",
    "pydub (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.14,<4.0"

[dependency-groups]
dev = [
    "ipython (>=9.11.0,<10.0.0)",
    "ipdb (>=0.13.13,<0.14.0)",
    "isort (>=8.0.1,<9.0.0)",
    "black (>=26.3.0,<27.0.0)"
]